+++ src/components/TripInfoCard.scss
@@ -1,0 +1,48 @@
+@import '../style/colors.scss';
+
+.TripInfoCard {
+  box-sizing: border-box;
+  width: 250px;
+  height: 332px;
+  margin: 0 0 40px 0;
+  border: 1px solid $grey200;
+  border-radius: 4px;
+  img {
+    width: 100%;
+    height: 187px;
+    border-top-right-radius: 3px;
+    border-top-left-radius: 3px;
+  }
+  &__info {
+    padding: 8px 16px;
+    &__tags {
+      margin: 4px 0 2px;
+      span {
+        font-size: 12px;
+        color: $grey500;
+        vertical-align: middle;
+        margin: 0 5px 0 0;
+      }
+    }
+    &__title {
+      width: 100%;
+      height: 45px;
+      margin: 0 0 20px;
+      font-size: 15px;
+      font-weight: 600;
+      line-height: 1.47;
+      color: $grey800;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      -webkit-box-orient: vertical;
+      display: -webkit-box;
+      -webkit-line-clamp: 2;
+    }
+    &__discription {
+      font-size: 14px;
+      color: $red400;
+      font-weight: 500;
+      vertical-align: middle;
+    }
+  }
+}
