+++ src/pages/Main/RecentTourList.scss
@@ -1,7 +1,7 @@
 @import "../../style/colors.scss";
-
-.MapSearchResult {
-  width: 530px;
+.RecentTourList {
+  width: 1060px;
+  margin: 80px auto 0;
   h2 {
     color: $grey700;
     font-size: 24px;
